Rename construction_slice.rs -> construction_view.rs
