+++ src/base/construction_view.rs
